requirements.txt
aiohttp==3.8.3
asyncpraw==7.6.1
discord.py==2.1.0
influxdb_client==1.35.0
psutil==5.9.4
python-dotenv==0.21.0
qrcode==7.3.1
